backend/requirements-metal.txt
# Metal GPU Acceleration Dependencies for M4 Max Optimization
# Nautilus Trading Platform - Apple Silicon Optimized
# Python 3.13.7 compatible versions
# Generated: 2025-08-24

# Core ML Acceleration Frameworks
# PyTorch with Metal Performance Shaders (MPS) support
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# Apple MLX Framework (Apple Silicon optimized)
mlx==0.28.0
mlx-lm==0.26.3
mlx-metal==0.28.0

# Core ML Tools
coremltools==8.3.0

# ML Acceleration and Optimization
accelerate==1.10.0
transformers==4.55.4
huggingface-hub==0.34.4
safetensors==0.6.2

# High-Performance Scientific Computing
numpy==2.3.2
scipy==1.16.1
pandas==2.3.1

# Performance Optimizations
bottleneck==1.5.0
numexpr==2.11.0

# High-Performance Data Processing
polars==1.32.3
pyarrow==21.0.0
fastparquet==2024.11.0
cramjam==2.11.0

# Quantitative Finance Libraries (Apple Silicon compatible)
ta-lib==0.6.5
quantlib-python==1.18
QuantLib==1.39

# Core Dependencies (from transformers/MLX)
tokenizers==0.21.4
regex==2025.7.34
hf-xet==1.1.8
cattrs==25.1.1
pyaml==25.7.0

# Build and Development Tools
build==1.3.0
pyproject_hooks==1.2.0

# Utilities and Support Libraries
filelock==3.13.1
fsspec==2025.7.0
packaging==25.0
setuptools==70.2.0
tqdm==4.67.1
psutil==7.0.0
PyYAML==6.0.2
requests==2.32.5

# Neural Network and Model Support
sympy==1.13.3
networkx==3.3
jinja2==3.1.4
MarkupSafe==2.1.5

# Compatibility Notes:
# - TensorFlow 2.20.0 has compatibility issues with macOS Sequoia 15.6
# - numba not compatible with Python 3.13 (upstream issue)
# - pyfolio and zipline-reloaded not compatible with Python 3.13
# - Use PyTorch MPS backend for GPU acceleration instead of CUDA
# - MLX provides optimal Apple Silicon performance for ML workloads

# Performance Benchmarks (M4 Max with 40 GPU cores):
# - Neural Network GPU Speedup: 4.89x over CPU
# - Matrix Operations: Variable speedup depending on workload size
# - Memory Usage: ~72MB GPU cache for typical ML models
# - Device: mps (Metal Performance Shaders)

# Installation Command:
# pip3 install -r requirements-metal.txt

# Verification:
# python3 -c "import torch; print('MPS available:', torch.backends.mps.is_available())"
# python3 -c "import mlx.core as mx; print('MLX device:', mx.default_device())"